+++ src/main/java/com/yuuko/core/database/GuildFunctions.java
@@ -166,6 +166,7 @@ public class GuildFunctions {
                 settings.add(rs.getBoolean("nowPlaying") ? "Enabled" : "Disabled");
                 settings.add(rs.getBoolean("djMode") ? "Enabled" : "Disabled");
                 settings.add(rs.getString("newMember"));
+                settings.add(rs.getString("newMemberMessage"));
                 settings.add(rs.getString("starboard"));
                 settings.add(rs.getString("commandLog"));
                 settings.add(rs.getString("modLog"));